refactor: Car이름 검증책임을 Car 도메인이 가지도록 변경
