+++ src/test/java/racingcar/CarTest.java
@@ -1,8 +1,10 @@
 package racingcar;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class CarTest {
     @Test
@@ -12,5 +14,22 @@ public class CarTest {
         assertThat(car.isLocated(1)).isTrue();
         car.move(3);
         assertThat(car.isLocated(1)).isTrue();
+    }
+
+    @Test
+    public void testNameIsNullOrEmpty(){
+        assertThatIllegalArgumentException().isThrownBy(()->{
+            new Car(null);
+        });
+        assertThatIllegalArgumentException().isThrownBy(()->{
+            new Car("");
+        });
+    }
+
+    @Test
+    public void testNameLengthIsOver5(){
+        assertThatIllegalArgumentException().isThrownBy(() -> {
+            new Car("carcar");
+        });
     }
 }
